replace folders in source code
