+++ pkg/log/impl/elk/kibana/client.go
