+++ pkg/log/impl/elk/kibana/model.go
