+++ pkg/log/impl/elk/result.go
