+++ pkg/log/impl/splunk/logclient/client.go
