+++ pkg/log/impl/splunk/logclient/client_test.go
