+++ pkg/log/impl/splunk/logclient/mapper.go
