+++ pkg/log/impl/splunk/logclient/mapper_test.go
