+++ pkg/log/impl/splunk/logclient/search.go
