+++ pkg/log/impl/splunk/main.go
